+++ src/main/java/net/frozenblock/lib/recipe/mixin/ShapelessRecipeBuilderMixin.java
@@ -19,12 +19,11 @@ package net.frozenblock.lib.recipe.mixin;
 
 import com.llamalad7.mixinextras.injector.wrapoperation.Operation;
 import com.llamalad7.mixinextras.injector.wrapoperation.WrapOperation;
-import net.frozenblock.lib.recipe.api.ShapedRecipeBuilderExtension;
+import net.frozenblock.lib.recipe.api.ShapelessRecipeBuilderExtension;
 import net.minecraft.advancements.AdvancementHolder;
 import net.minecraft.core.component.DataComponentPatch;
 import net.minecraft.data.recipes.RecipeOutput;
-import net.minecraft.data.recipes.ShapedRecipeBuilder;
-import net.minecraft.resources.ResourceKey;
+import net.minecraft.data.recipes.ShapelessRecipeBuilder;
 import net.minecraft.resources.ResourceLocation;
 import net.minecraft.world.item.crafting.Recipe;
 import net.minecraft.world.item.crafting.ShapedRecipe;
@@ -33,8 +32,8 @@ import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.Unique;
 import org.spongepowered.asm.mixin.injection.At;
 
-@Mixin(ShapedRecipeBuilder.class)
-public class ShapedRecipeBuilderMixin implements ShapedRecipeBuilderExtension {
+@Mixin(ShapelessRecipeBuilder.class)
+public class ShapelessRecipeBuilderMixin implements ShapelessRecipeBuilderExtension {
 
 	@Unique
 	@Nullable
@@ -42,20 +41,20 @@ public class ShapedRecipeBuilderMixin implements ShapedRecipeBuilderExtension {
 
 	@Unique
 	@Override
-	public ShapedRecipeBuilder frozenLib$patch(@Nullable DataComponentPatch patch) {
+	public ShapelessRecipeBuilder frozenLib$patch(@Nullable DataComponentPatch patch) {
 		this.patch = patch;
-		return (ShapedRecipeBuilder) (Object) this;
+		return (ShapelessRecipeBuilder) (Object) this;
 	}
 
 	@WrapOperation(
 		method = "save",
 		at = @At(
 			value = "INVOKE",
-			target = "Lnet/minecraft/data/recipes/RecipeOutput;accept(Lnet/minecraft/resources/ResourceKey;Lnet/minecraft/world/item/crafting/Recipe;Lnet/minecraft/advancements/AdvancementHolder;)V"
+			target = "Lnet/minecraft/data/recipes/RecipeOutput;accept(Lnet/minecraft/resources/ResourceLocation;Lnet/minecraft/world/item/crafting/Recipe;Lnet/minecraft/advancements/AdvancementHolder;)V"
 		)
 	)
-	private void modifySave(RecipeOutput instance, ResourceKey<Recipe<?>> recipeId, Recipe<?> recipe, AdvancementHolder holder, Operation<ShapedRecipe> operation) {
-		((ShapedRecipeBuilderExtension) recipe).frozenLib$patch(this.patch);
+	private void modifySave(RecipeOutput instance, ResourceLocation recipeId, Recipe<?> recipe, AdvancementHolder holder, Operation<ShapedRecipe> operation) {
+		((ShapelessRecipeBuilderExtension) recipe).frozenLib$patch(this.patch);
 		operation.call(instance, recipeId, recipe, holder);
 	}
 }